New kernel and EE7 images.
Previous latest image now tagged webProfile6. New webProfile7 image added that contains Java EE7 Web Profile features. The webProfile6 and webProfile7 images build on a common image which provides a base set of features. This in turn builds on a kernel image that contains no runtime features that can be used as the basis for custom images. Also added is a javaee7 image which builds on webProfile7 and adds all of the remaining features for Java EE7 Full Platform.

The image tags with version numbers only exist to signpost the current version being used. Liberty has a 'zero migration' policy which means a given feature set should continue to work regardless of the version number.