[security] Update for Node.js v7.2.0
This update includes is a security fix for v7 (v7.2.0) that fixes a
potential buffer overflow when writing data to console on Windows 10.

This update also includes the addition of a node group/user with
gid/uid 1000, a home directory and sets bash as the default shell for
the node user (except the Alpine variant).

Reference:

- https://nodejs.org/en/blog/release/v7.2.0/
- https://github.com/nodejs/docker-node/pull/274
- https://github.com/nodejs/docker-node/pull/275
- https://github.com/nodejs/docker-node/issues/273
- https://github.com/nodejs/docker-node/pull/263