telegraf 1.3.0 and smaller chronograf image
Update chronograf, influxdb, kapacitor, and telegraf so they check
multiple gpg servers for a key rather than failing if the keyserver
fails. This should make the builds more reliable.

chronograf is updated to use jessie-slim so the image will be smaller.
If this works out well, we'll be following the same pattern with the
other images. This also updates the license documentation within the
chronograf image.