Update Amazon Linux images
Amazon Linux 2 is available and is out of the candidate stage. We
dropped the YYYY.MM versioning scheme in this release. "latest" now
points to AL2.

We've updated the container image for 2018.03.

I've dropped references to all other versions, as we no longer generally
support those (but as I understand it, this does not keep users from
pulling those images as tags never get deleted from Docker Hub).