Add haskell-stack test, update haskell-cabal test.
Stack is a tool we've included in the haskell image since 7.10.3, so it only
makes sense to add a test for it. In addition, the updates to the haskell-cabal
test clean up the run.sh output and validate basic installation of a package.
Both the stack and cabal tests download (different) upstream package indexes,
so they can take more than a few seconds each to complete.