Important bug fix to 4.0.0 authentication
There was a bug in the `NEO4J_AUTH` logic that meant the default password wasn't being set if `NEO4J_AUTH` was empty. This is fixed, and I feel is sufficiently important that we should update 4.0.0 tag (the actual neo4j source stays the same, just the docker-entrypoint wrapper is changed).