consul: switch generic beta tag to 1.8.0-beta
The generic `beta` tag was unclear and would be ephemeral (appearing shortly before new releases rather than being a consistent channel like `latest` or `nightly`). This switches to a more specific `1.8.0-beta` tag to be more descriptive and allow clearly opting in to testing a series of beta releases for a specific upcoming release.