Adjust nextcloud tests to avoid flakiness
This adjusts `nextcloud-apache-run` and `nextcloud-fpm-run` to try up to three times for a successful result, and adds a new `nextcloud-cli` test that combines the other three (`nextcloud-cli-mysql`, `nextcloud-cli-postgres`, and `nextcloud-cli-sqlite`) into a single test that will pass if any one of them passes (but does attempt to run all three every time).

Also, account for change in the installation page (from "Finish setup" to "Install").