Update external-pins diffing to match normal diffing
When we run `diff-pr.sh` on PRs, it normally shows us a diff of either `amd64` or the first supported architecture.

However, when we implemented support for external pins, I was very silly and had it show us _all_ architectures.

This makes for very long diffs.

In this change, I've updated that external pins logic to match the prior logic -- in lieu of better diffing algorithms (targeting review instead of patching) that can do smarter things like collapsing deltas that are identical, I think this is the right thing to do.