Re-enable BuildKit "Garbage Collection"
Initially, I had high hopes that we would be able to manage BuildKit's cache in similar ways to Docker's, especially because we happen to be able to know exactly the set of built images for whom cache should be kept (and everything else is ripe for deletion).  In practice however, this is far too cumbersome to manage correctly, especially as we have had to dramatically change the way we perform these builds over time such that this is no longer reasonable.

In response to that, this adjusts our BuildKit configuration to no longer disable Garbage Collection such that it should automatically clean up what it determines to be "stale" according to built-in heuristics. 🤷