MariaDB 2024 Q2 part 1 release
The new 11.4 release is a little late going though QA so it will
occur later. In the mean time we keep the 11.3 as the latest tag
so it isn't a downgrade.

As such 11.3 and 11.4 don't have the fix for issue 560.

We also filter our spaces in the command when passing to the
mariadb-install-db as it cannot handle this currently.