Reinstate Kong Alpine 2.8.5 as the main version
There was some confusion during the release of kong 2.8.5.
While on the 3.x series we no longer maintain Alpine, it is still
maintained in 2.x

As such, this PR reinstates Alpine as the head for 2.8.x.

The change in the Ubuntu SHA is due to the fact that we had to add a new
commit to the branch in order to rebuild alpine. The ubuntu files are
otherwise unchanged